dashboard + mixins
