+++ backend/uofcPortal/api/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from .views import UserViewSet, StudentViewSet, FacultyViewSet, DepartmentViewSet, ProgramViewSet, PersonalInfoViewSet, CourseViewSet, InstructorViewSet, LectureViewSet, GradeViewSet, EnrollmentViewSet, StudentGradeView, StudentFinancesView, TransactionViewSet
+from .views import UserViewSet, StudentViewSet, FacultyViewSet, DepartmentViewSet, ProgramViewSet, PersonalInfoViewSet, CourseViewSet, InstructorViewSet, LectureViewSet, GradeViewSet, EnrollmentViewSet, StudentGradeView, StudentFinancesView, TransactionViewSet, StudentApplicationsViewSet, DashboardView
 
 router = DefaultRouter()
 router.register('users', UserViewSet)
@@ -15,12 +15,11 @@ router.register('lectures', LectureViewSet, basename='lecture')
 router.register('grades', GradeViewSet, basename='grade')
 router.register('enrollments', EnrollmentViewSet, basename='enrollment')
 router.register('transactions', TransactionViewSet, basename='transaction')
-
+router.register('student-applications', StudentApplicationsViewSet, basename='student-applications') # To be modifed
 
 urlpatterns = [
     path('api/', include(router.urls)),
     path('api/student-grades/', StudentGradeView.as_view(), name='student-grade'),
     path('api/student-finances/', StudentFinancesView.as_view(), name='student-finances'),
-
-
+    path('api/dashboard/', DashboardView.as_view(), name='dashboard'),
 ]
